Final changes as per feedback
