+++ apps/quick-learn-backend/src/routes/metadata/metadata.module.ts
@@ -3,6 +3,8 @@ import { MetadataController } from './metadata.controller';
 import { MetadataService } from './metadata.service';
 import { RoadmapCategoriesModule } from '../roadmap-category/roadmap-category.module';
 import { CourseCategoriesModule } from '../course-category/course-category.module';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { CourseCategoryEntity, RoadmapCategoryEntity } from '@src/entities';
 
 @Module({
   imports: [RoadmapCategoriesModule, CourseCategoriesModule],
