+++ apps/quick-learn-backend/src/routes/metadata/metadata.service.ts
@@ -1,53 +1,20 @@
 import { Injectable } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { CourseCategoryEntity, RoadmapCategoryEntity } from '@src/entities';
-import { Repository } from 'typeorm';
+import { RoadmapCategoryService } from '../roadmap-category/roadmap-category.service';
+import { CourseCategoryService } from '../course-category/course-category.service';
 
 @Injectable()
 export class MetadataService {
   constructor(
-    @InjectRepository(RoadmapCategoryEntity)
-    @InjectRepository(CourseCategoryEntity)
-    private roadmapCategoryRepository: Repository<RoadmapCategoryEntity>,
-    private courseCategoryRepository: Repository<CourseCategoryEntity>,
+    private readonly roadmapCategoryService: RoadmapCategoryService,
+    private readonly courseCategoryService: CourseCategoryService,
   ) {}
 
   async getContentRepositoryMetadata() {
     const metadata = {};
     metadata['roadmap_categories'] =
-      await this.getRoadmapCatergoriesWithRoadmap();
-    metadata['course_categories'] = await this.getCourseCategoriesWithCourses();
+      await this.roadmapCategoryService.getRoadmapCatergoriesWithRoadmap();
+    metadata['course_categories'] =
+      await this.courseCategoryService.getCourseCategoriesWithCourses();
     return metadata;
-  }
-
-  async getRoadmapCatergoriesWithRoadmap() {
-    return await this.roadmapCategoryRepository
-      .createQueryBuilder('category')
-      .leftJoinAndSelect('category.roadmaps', 'roadmaps')
-      .where('roadmaps.archived = :archived', { archived: false })
-      .orderBy('category.name', 'ASC')
-      .addOrderBy('category.created_at', 'DESC')
-      .getMany();
-  }
-  async getCourseCategoriesWithCourses() {
-    return await this.courseCategoryRepository
-      .createQueryBuilder('course_category')
-      .leftJoinAndSelect(
-        'course_category.courses',
-        'courses',
-        'courses.archived = :archived',
-        { archived: false },
-      )
-      .leftJoin('courses.lessons', 'lessons')
-      .loadRelationCountAndMap(
-        'courses.lessons_count',
-        'courses.lessons',
-        'lessons',
-        (qb) =>
-          qb.andWhere('lessons.archived = :archived', { archived: false }),
-      )
-      .orderBy('courses.created_at', 'DESC')
-      .addOrderBy('course_category.created_at', 'DESC')
-      .getMany();
   }
 }
